WIP work for idle animations.
Note: the current problem is that the large images take
too long to load, and load certain layers before others.

However, once loaded the images move smoothly, so one 
possible solution is to load the static image first, load
the larger animatable image outside the visible area of the
div, and then slide the animatable image in when ready.