Baseline: add ability to jump straight into task
Using the task route ?baseline/x/y, we can jump
straight into task x/y without ever seeing the world
or the town.

Added baselineController as basic flag for views or
actions that need to branch based on whether we are
baseline or not. Hopefully we don't need to do too
much of this -- if we do we should think of another
architecture or make a branch of the codebase to
prevent too much branching in the code.