Add meiosisControl prop to task; reflect in switch behavior
If meiosisControl is false, switch will not flip and tooltip
will reflect that (eventually we probably want to change the
graphic as well).

Currently if it is any other value (default is "all"),
meiosis will be allowed and will continue as normal. 
Next we will make different behaviors for different values.