Add ability to swap between arbitrary animations.
Right now, we are just swapping between headTurn and
the shine animations.

The code now swaps in the static image between each
animation and waits for the newest images to be loaded
before trying to resize and then animate them.

Some of the waiting logic was trimmed down. We no longer
create a fake image and wait for it to download, we simply
use the existing imagesLoaded library. Also one full
setTimeout was removed as being unneeded, but the first one
(between imagesLoaded saying the images are ready and
actually displaying the images) was made slightly longer
to compensate. Over all, though, users wait less time
between animations now.