Switch to baseline world if using baseline route
Routes starting with ?task=baseline/ switch to baseline
mode as usual. Now this also causes the towns to be
loaded from the "baseline" world instead of the "game"
world.

Update the local version of api/game to the latest
version from dev, which includes a baseline world with
four out of six towns authored.