Add connection-lost notification in tournament worlds
If we fail to save once, then fail to empty the queue once,
we send a notification to the statechart that our connection
was lost.

If we are in a tournament world, the statechart will then 
pop up a large notification to the user.

In the future, we could extend this to a small warning in 
non-tournament world.

When the connection is regained and the data is saved we 
briefly notify the user.

[#50512117]