Add manual event submission panel for tournament world
After logging in as a user, an event organizer can click on
the word "Hi" in the top bar and enter the password
"gen1games".

This will bring up an event submission dialog box, which
will allow the user to submit both a drake submission event
and a task-end event. Both these events must always be sent
together. If specific drake alleles are not sent, a dummy
drake is sent instead.

Currently, this dialog is only used when we want to submit
task-end events (with/without accompanying drakes), as the 
tournament doesn't care about other events.