Enable towns based on completion; unlock towns manually
This reverts town-enabling back to how it was before. We now
have a separate "locked" state that is represented in the UI,
can be unlocked, and is saved to the backend.

Also update log messages to log unlocking event.

Update stylesheet to allow completed but not enabled towns. (Enabled
has become a bit of a misnomer here.)

[#59758674]