Add animation for changing score
NB: This is done using a computed property that observes the 
real property and changes slowly to meet it. This is probably
not the right way to do this as it is simply a view-level
feature, but it still seems most ember-like for the view
to bind to a real property, instead of using an animation
to pick what number should be shown in the score field.

[#60200496]