Add annoying prop to breedingController so we know if we're showing breeder
We want the GoToTownButton view to easily bind to GG.breedingController.isShowingBreeder,
but because of the removeObserver call on this property in DrakeView, the binding gets
destroyed. Changing how we add and remove the observer in the DrakeView does not
appear to fix this problem. So this is a hack, adding another property that only
the button observes...

[#44727887]