Create trait-selection task-header with radio-buttons
This adds a selection panel on top of the breeder panel
which allows students to select the characteristics they've
been asked to breed.

There are two separate panels, one is used for towns 3+, when
there are more traits.

Clicking a radio button simply lights it up and logs the
selection.

[#69207882]