Move back to town on submit; show failure message on fail
We now move back to town when a drake has been submitted in
all cases. If the drake is not correct, we show a failure message,
either authored (npc.speech.failText) or predefined. After the
failure message, we give the option to repeat task or to dismiss
the speech bubble.

On restart, correctly log restart to the user's log.

[#69207464]