Show Breeders License message; add Pay buttons if user can afford
When we show the town unlock dialog, we now show a message with
the Breeders License and the cost required to unlock the town.

If the user cannot afford the unlock cost (*or* if the user
is trying to skip to a later town), we show the regular password
field.

[#70312444]