Add registration form, allow registration for new users
This submits the registration form to the portal, which returns a success
message which we parse to grab the student's new ID.

We pre-fill-in the hidden classname field with a default class name,
in this case default-class. This class is owned by teacher.

Currently, success is assumed. We should see if we need to check
this assumption.

After registration, the username is displayed and the user can
log in.

[#70470592]